+++ h2o-algos/src/test/java/hex/pdp/PartialDependenceTest.java
@@ -1,6 +1,6 @@
 package hex.pdp;
 
-import hex.PDP;
+import hex.PartialDependence;
 import hex.tree.gbm.GBM;
 import hex.tree.gbm.GBMModel;
 import org.junit.Assert;
@@ -14,13 +14,13 @@ import water.fvec.Vec;
 import water.util.Log;
 import water.util.TwoDimTable;
 
-public class PDPTest extends TestUtil {
+public class PartialDependenceTest extends TestUtil {
   @BeforeClass() public static void setup() { stall_till_cloudsize(1); }
 
   @Test public void prostateBinary() {
     Frame fr=null;
     GBMModel model=null;
-    PDP pdp = null;
+    PartialDependence partialDependence = null;
     try {
       // Frame
       fr = parse_test_file("smalldata/prostate/prostate.csv");
@@ -38,28 +38,28 @@ public class PDPTest extends TestUtil {
       parms._response_column = "CAPSULE";
       model = new GBM(parms).trainModel().get();
 
-      // PDP
-      pdp = new PDP(Key.<PDP>make());
-//      pdp._cols = model._output._names;
-      pdp._nbins = 10;
-      pdp._model_id = (Key) model._key;
-      pdp._frame_id = fr._key;
+      // PartialDependence
+      partialDependence = new PartialDependence(Key.<PartialDependence>make());
+//      partialDependence._cols = model._output._names;
+      partialDependence._nbins = 10;
+      partialDependence._model_id = (Key) model._key;
+      partialDependence._frame_id = fr._key;
 
-      pdp.execImpl().get();
-      for (TwoDimTable t : pdp._partial_dependence_data)
+      partialDependence.execImpl().get();
+      for (TwoDimTable t : partialDependence._partial_dependence_data)
         Log.info(t);
 
     } finally {
       if (fr!=null) fr.remove();
       if (model!=null) model.remove();
-      if (pdp!=null) pdp.remove();
+      if (partialDependence !=null) partialDependence.remove();
     }
   }
 
   @Test public void prostateBinaryPickCols() {
     Frame fr=null;
     GBMModel model=null;
-    PDP pdp = null;
+    PartialDependence partialDependence = null;
     try {
       // Frame
       fr = parse_test_file("smalldata/prostate/prostate.csv");
@@ -77,30 +77,30 @@ public class PDPTest extends TestUtil {
       parms._response_column = "CAPSULE";
       model = new GBM(parms).trainModel().get();
 
-      // PDP
-      pdp = new PDP(Key.<PDP>make());
-      pdp._cols = new String[]{"DPROS", "GLEASON"}; //pick columns manually
-      pdp._nbins = 10;
-      pdp._model_id = (Key) model._key;
-      pdp._frame_id = fr._key;
+      // PartialDependence
+      partialDependence = new PartialDependence(Key.<PartialDependence>make());
+      partialDependence._cols = new String[]{"DPROS", "GLEASON"}; //pick columns manually
+      partialDependence._nbins = 10;
+      partialDependence._model_id = (Key) model._key;
+      partialDependence._frame_id = fr._key;
 
-      pdp.execImpl().get();
-      for (TwoDimTable t : pdp._partial_dependence_data)
+      partialDependence.execImpl().get();
+      for (TwoDimTable t : partialDependence._partial_dependence_data)
         Log.info(t);
 
-      Assert.assertTrue(pdp._partial_dependence_data.length == 2);
+      Assert.assertTrue(partialDependence._partial_dependence_data.length == 2);
 
     } finally {
       if (fr!=null) fr.remove();
       if (model!=null) model.remove();
-      if (pdp!=null) pdp.remove();
+      if (partialDependence !=null) partialDependence.remove();
     }
   }
 
   @Test public void prostateRegression() {
     Frame fr=null;
     GBMModel model=null;
-    PDP pdp = null;
+    PartialDependence partialDependence = null;
     try {
       // Frame
       fr = parse_test_file("smalldata/prostate/prostate.csv");
@@ -118,27 +118,27 @@ public class PDPTest extends TestUtil {
       parms._response_column = "AGE";
       model = new GBM(parms).trainModel().get();
 
-      // PDP
-      pdp = new PDP(Key.<PDP>make());
-      pdp._nbins = 10;
-      pdp._model_id = (Key) model._key;
-      pdp._frame_id = fr._key;
+      // PartialDependence
+      partialDependence = new PartialDependence(Key.<PartialDependence>make());
+      partialDependence._nbins = 10;
+      partialDependence._model_id = (Key) model._key;
+      partialDependence._frame_id = fr._key;
 
-      pdp.execImpl().get();
-      for (TwoDimTable t : pdp._partial_dependence_data)
+      partialDependence.execImpl().get();
+      for (TwoDimTable t : partialDependence._partial_dependence_data)
         Log.info(t);
 
     } finally {
       if (fr!=null) fr.remove();
       if (model!=null) model.remove();
-      if (pdp!=null) pdp.remove();
+      if (partialDependence !=null) partialDependence.remove();
     }
   }
 
   @Test public void weatherBinary() {
     Frame fr=null;
     GBMModel model=null;
-    PDP pdp = null;
+    PartialDependence partialDependence = null;
     try {
       // Frame
         fr = parse_test_file("smalldata/junit/weather.csv");
@@ -150,21 +150,21 @@ public class PDPTest extends TestUtil {
       parms._response_column = "RainTomorrow";
       model = new GBM(parms).trainModel().get();
 
-      // PDP
-      pdp = new PDP(Key.<PDP>make());
-      pdp._nbins = 33;
-      pdp._cols = new String[]{"Sunshine","MaxWindPeriod","WindSpeed9am"};
-      pdp._model_id = (Key) model._key;
-      pdp._frame_id = fr._key;
+      // PartialDependence
+      partialDependence = new PartialDependence(Key.<PartialDependence>make());
+      partialDependence._nbins = 33;
+      partialDependence._cols = new String[]{"Sunshine","MaxWindPeriod","WindSpeed9am"};
+      partialDependence._model_id = (Key) model._key;
+      partialDependence._frame_id = fr._key;
 
-      pdp.execImpl().get();
-      for (TwoDimTable t : pdp._partial_dependence_data)
+      partialDependence.execImpl().get();
+      for (TwoDimTable t : partialDependence._partial_dependence_data)
         Log.info(t);
 
     } finally {
       if (fr!=null) fr.remove();
       if (model!=null) model.remove();
-      if (pdp!=null) pdp.remove();
+      if (partialDependence !=null) partialDependence.remove();
     }
   }
