+++ h2o-core/src/main/java/water/api/ModelsHandler.java
@@ -5,7 +5,7 @@ import java.net.URI;
 import java.util.*;
 
 import hex.Model;
-import hex.PDP;
+import hex.PartialDependence;
 import water.*;
 import water.api.FramesHandler.Frames;
 import water.api.schemas3.*;
@@ -163,20 +163,20 @@ public class ModelsHandler<I extends ModelsHandler.Models, S extends SchemaV3<I,
   }
 
   @SuppressWarnings("unused") // called from the RequestServer through reflection
-  public JobV3 makePDP(int version, PDPV3 s) {
-    PDP pdp;
+  public JobV3 makePartialDependence(int version, PartialDependenceV3 s) {
+    PartialDependence partialDependence;
     if (s.destination_key != null)
-      pdp = new PDP(s.destination_key.key());
+      partialDependence = new PartialDependence(s.destination_key.key());
     else
-      pdp = new PDP(Key.<PDP>make());
-    s.fillImpl(pdp); //fill frame_id/model_id/nbins/etc.
-    return new JobV3(pdp.execImpl());
+      partialDependence = new PartialDependence(Key.<PartialDependence>make());
+    s.fillImpl(partialDependence); //fill frame_id/model_id/nbins/etc.
+    return new JobV3(partialDependence.execImpl());
   }
 
   @SuppressWarnings("unused") // called from the RequestServer through reflection
-  public PDPV3 fetchPDP(int version, KeyV3.PDPKeyV3 s) {
-    PDP pdp = DKV.getGet(s.key());
-    return new PDPV3().fillFromImpl(pdp);
+  public PartialDependenceV3 fetchPartialDependence(int version, KeyV3.PartialDependenceKeyV3 s) {
+    PartialDependence partialDependence = DKV.getGet(s.key());
+    return new PartialDependenceV3().fillFromImpl(partialDependence);
   }
 
   /** Remove an unlocked model.  Fails if model is in-use. */
